Master Sync

+++ Server/Settings/Definition/LogSettingsDefinition.cs
@@ -11,7 +11,7 @@ namespace Server.Settings.Definition
         public LogLevels LogLevel { get; set; } = LogLevels.Debug;
 
         [XmlComment(Value = "Specify the amount of days a log file should be considered as expired and deleted. 0 = Disabled")]
-        public double ExpireLogs { get; set; } = 14;
+        public int ExpireLogs { get; set; } = 14;
 
         [XmlComment(Value = "Use UTC instead of system time in the log.")]
         public bool UseUtcTimeInLog { get; set; } = false;
